+++ packages/fepanel/main.js
@@ -47,7 +47,7 @@ const plugin = {
     const resolver = function (host, url, req) {
       const formattedPath = `${host}${url}`;
       const formattedDomain = opts.domain.replace(/\./g, '\\.');
-      const staticFilesTester = new RegExp(`^${formattedDomain}\/(css|fonts|js)\/.+$`);
+      const staticFilesTester = new RegExp(`^${formattedDomain}\/((css|fonts|js)\/.+)|(.+\.(js|ico))$`);
       const domainTester = new RegExp(`^${formattedDomain}\/`);
       if (staticFilesTester.test(formattedPath)) {
         return `http://127.0.0.1:${app.tigo.config.server.port}/static/fepanel/`;